Fix : board 삭제 로직 수정(Matching과 MatchingApply가 null이 아닌 경우에) + matchin…

+++ src/main/java/com/back/moment/boards/service/BoardService.java
@@ -175,9 +175,12 @@ public class BoardService {
             throw new ApiException(ExceptionEnum.NOT_MATCH_USERS);
         }
         List<MatchingApply> matchingApplyList = matchingApplyRepository.findAllByBoardId(boardId);
-        matchingApplyRepository.deleteAll(matchingApplyList);
+        if(matchingApplyList != null)
+            matchingApplyRepository.deleteAll(matchingApplyList);
+
         Matching matching = matchingRepository.findByBoardId(boardId);
-        matchingRepository.delete(matching);
+        if(matching != null)
+            matchingRepository.delete(matching);
 
         s3Uploader.delete(board.getBoardImgUrl());
         boardRepository.deleteById(boardId);
